Modify query by country

+++ crawer.py
@@ -19,7 +19,7 @@ class Crawler:
             month=self.month
         )
         self.dirname = Template(
-            "$path/$sid/$year/$month/"
+            "$path/data/$sid/$year/"
         ).substitute(**self.args)
         self.args['dirname'] = self.dirname
         self.filename = Template(
@@ -31,9 +31,6 @@ class Crawler:
         self.params_url = Template(
             "$burl?yrmth_crit=$year-$month&graph_category=1&gesendet=Start+search&loggerid_crit=$sid"
         ).substitute(**self.args)
-
-        self.__make_directories()
-        self.__fetch_and_save_data()
 
     def __make_directories(self):
         if not os.path.exists(self.dirname):
@@ -55,39 +52,70 @@ class Crawler:
             print(f"[{res.status_code}] request failed!")
             return None
 
-    def __fetch_and_save_data(self):
-        table = self.__request_data()
-        print(table)
-        if table is None:
-            print("Data not found")
-            return
+    def fetch_and_save_data(self):
+        if os.path.exists(self.filename):
+            print('File was exist...')
+        else:
+            self.__make_directories()
 
-        data = dict()
-        for i, row in enumerate(table.find_all('tr')):
-            if i == 1:
-                continue
-            if i == 0:
-                for j, col in enumerate(row.find_all('td')):
-                    col = col.get_text()
-                    data[col] = list()
-            else:
-                col_values = row.find_all('td')
-                for j, col_names in enumerate(data.keys()):
-                    value = col_values[j].get_text().replace('\n', '')
-                    data[col_names].append(value)
+            table = self.__request_data()
+            if table is None:
+                print("Data not found")
+                return
+
+            data = dict()
+            for i, row in enumerate(table.find_all('tr')):
+                if i == 1:
+                    continue
+                if i == 0:
+                    for j, col in enumerate(row.find_all('td')):
+                        col = col.get_text()
+                        data[col] = list()
+                else:
+                    col_values = row.find_all('td')
+                    for j, col_names in enumerate(data.keys()):
+                        value = col_values[j].get_text().replace('\n', '')
+                        data[col_names].append(value)
 
 
-        data = pd.DataFrame(data)
+            data = pd.DataFrame(data)
 
-        print(data)
+            print(data)
 
-        data.to_csv(self.filename, header=True, index=False, sep='\t', mode='w')
-        print(f"File saved in {self.filename}...")
+            data.to_csv(self.filename, header=True, index=False, sep='\t', mode='w')
+            print(f"File saved in {self.filename}...")
 
 
 if __name__ == "__main__":
-    crawer = Crawler(station_id=857341, year=2014, month=10)
-    crawer
+    PATH = "./input_data/Namibia"
+    freq = 'DAILYURL'
+    all_files = list()
+
+    for root, dir, files in os.walk(PATH):
+        for file in files:
+            all_files.append(os.path.join(root, file))
+
+    all_files = list(
+        filter(
+            lambda file: freq in file, all_files
+        )
+    )
+
+    for file in all_files:
+        print("file: ", file)
+        with open(file, 'r') as handle:
+            params = list(
+                dict(
+                    station_id=row.split()[0],
+                    year=row.split()[1],
+                    month=row.split()[2]
+                ) for row in handle.read().strip().split('\n')
+            )
+        total = len(params)
+        for i, param in enumerate(params):
+            print("Parameter %d/%d: " % (i+1, total), param)
+            crawer = Crawler(**param)
+            crawer.fetch_and_save_data()
